requirements.txt
pillow
transformers
torch
googletrans==4.0.0-rc1 